refactor: rename DashboardToolbar -> Toolbar
