+++ packages/ui/src/dashboard/index.ts
@@ -1,4 +1,4 @@
 export * from "./Statistics";
 export * from "./Storage";
 export * from "./Layout";
-export * from "./DashboardToolbar";
+export * from "./Toolbar";
